Adds `Tensor::get_array` and `Tensor::get_array_mut`
Note: because the inner struct for `Tensor` is an `Rc<TensorCore>`,
`Tensor::get_array_mut` will only provide the `&mut` reference to the
backing array if the `Tensor` has not been copied. This is a limitation
of how `Rc` works. More generally, `Rc` is immutable except if the
wrapped type has interior mutability (`Mutex`, `RefCell`, etc.). A
`Rc<RefCell<TensorCore>>` may work better in this situation.